# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bq — Birthday Queue",
    "version": "0.1.0"
  },
  "paths": {
    "/api/session": {
      "post": {
        "tags": [
          "convidado"
        ],
        "summary": "Create Session",
        "description": "Cria ou reidentifica o convidado. Idempotente por cookie: com cookie válido isto é\n`UPDATE` do MESMO convidado, nunca um segundo.",
        "operationId": "create_session_api_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "convidado"
        ],
        "summary": "Rename Session",
        "description": "RF-03. Renomeia **o mesmo** convidado: `UPDATE`, nunca `INSERT`.\n\n🔴 Se isto criasse um convidado novo, o cooldown de RF-09 morreria — e não por má fé: a\nprimeira pessoa que trocasse o apelido descobriria por acidente que o cooldown zerou, e\ncontaria para as outras. É a única defesa de cota que sobrou depois do corte de segurança,\ne ela cabe na escolha entre dois verbos SQL.",
        "operationId": "rename_session_api_session_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/suggestions": {
      "post": {
        "tags": [
          "convidado"
        ],
        "summary": "Suggest",
        "description": "A ORDEM DE VALIDAÇÃO É NORMATIVA (05 §3), porque decide qual mensagem a pessoa vê.\n\nE o cooldown é verificado no passo 2 mas gravado no passo 7: assim uma tentativa recusada\nnão gasta a vez (RF-09) — quem escolheu uma música de 9 minutos e levou `TOO_LONG` escolhe\noutra imediatamente, em vez de esperar 2 minutos por um erro.",
        "operationId": "suggest_api_suggestions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/suggestions/{suggestion_id}": {
      "delete": {
        "tags": [
          "convidado"
        ],
        "summary": "Remove Suggestion",
        "description": "RF-14. Só a própria, só enquanto não começou a tocar.\n\n**Não devolve a cota do cooldown**, e isso é a outra metade da mesma decisão: sem ela,\n\"sugerir e remover\" seria um jeito acidental de manter a fila inteira sob controle de uma\npessoa, e alguém descobre isso sem querer nos primeiros 20 minutos.",
        "operationId": "remove_suggestion_api_suggestions__suggestion_id__delete",
        "parameters": [
          {
            "name": "suggestion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Suggestion Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/skip-votes": {
      "post": {
        "tags": [
          "convidado"
        ],
        "summary": "Cast Vote",
        "operationId": "cast_vote_api_skip_votes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "convidado"
        ],
        "summary": "Retract Vote",
        "operationId": "retract_vote_api_skip_votes_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionIn": {
        "properties": {
          "nickname": {
            "type": "string",
            "title": "Nickname"
          }
        },
        "type": "object",
        "required": [
          "nickname"
        ],
        "title": "SessionIn"
      },
      "SessionOut": {
        "properties": {
          "guestId": {
            "type": "integer",
            "title": "Guestid"
          },
          "nickname": {
            "type": "string",
            "title": "Nickname"
          },
          "cooldownUntilMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cooldownuntilms"
          }
        },
        "type": "object",
        "required": [
          "guestId",
          "nickname",
          "cooldownUntilMs"
        ],
        "title": "SessionOut"
      },
...
      "SuggestIn": {
        "properties": {
          "trackId": {
            "type": "string",
            "title": "Trackid"
          }
        },
        "type": "object",
        "required": [
          "trackId"
        ],
        "title": "SuggestIn"
      },
      "SuggestOut": {
        "properties": {
          "suggestionId": {
            "type": "integer",
            "title": "Suggestionid"
          },
          "positionHint": {
            "type": "string",
            "title": "Positionhint"
          },
          "cooldownUntilMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cooldownuntilms"
          }
        },
        "type": "object",
        "required": [
          "suggestionId",
          "positionHint",
          "cooldownUntilMs"
        ],
        "title": "SuggestOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VoteIn": {
        "properties": {
          "playId": {
            "type": "integer",
            "title": "Playid"
          }
        },
        "type": "object",
        "required": [
          "playId"
        ],
        "title": "VoteIn"
      },
      "VoteOut": {
        "properties": {
          "votes": {
            "type": "integer",
            "title": "Votes"
          },
          "needed": {
            "type": "integer",
            "title": "Needed"
          },
          "youVoted": {
            "type": "boolean",
            "title": "Youvoted"
          }
        },
        "type": "object",
        "required": [
          "votes",
          "needed",
          "youVoted"
        ],
        "title": "VoteOut"
      },
...
    }
  }
}